Issues in Communication module
Babar please properly test and check in again. New changes are not working.

+++ KUPFApi/API/Servivces/Implementation/DetailedEmployee/DetailedEmployeeService.cs
@@ -65,7 +65,7 @@ namespace API.Servivces.Implementation.DetailedEmployee
                 var newEmployee = _mapper.Map<Models.DetailedEmployee>(detailedEmployeeDto);
                 newEmployee.LocationId = 1;
                 newEmployee.CRUP_ID = maxCrupId;
-                newEmployee.EmployeeId = CommonMethods.CreateEmployeeId();
+                newEmployee.EmployeeId = (int)CommonMethods.CreateEmployeeId();
                 await _context.DetailedEmployees.AddAsync(newEmployee);
                 await _context.SaveChangesAsync();
                 //
